fix: gRPC port

+++ util/port.go
@@ -13,7 +13,7 @@ import (
 // GetPortAvailBetween 获取10000到10100之间的随机整数
 func GetPortAvailBetween() (uint, error) {
 	rand.Seed(time.Now().Unix())
-	port := 10000
+	port := uint(10000)
 	for !IsPortAvailable(port) {
 		port = port + 1
 		if port > 10100 {
@@ -24,7 +24,7 @@ func GetPortAvailBetween() (uint, error) {
 }
 
 // IsPortAvailable 判断端口是否可用
-func IsPortAvailable(port int) bool {
+func IsPortAvailable(port uint) bool {
 	address := fmt.Sprintf("127.0.0.1:%d", port)
 	listener, err := net.Listen("tcp", address)
 	if err != nil {
